+++ packages/bech32m/src/format/prefix.ts
@@ -2,13 +2,14 @@ export const Prefixes = {
   passet: 'passet',
   pauctid: 'pauctid',
   penumbra: 'penumbra',
-  penumbracompat1: 'penumbracompat1',
   penumbrafullviewingkey: 'penumbrafullviewingkey',
   penumbragovern: 'penumbragovern',
   penumbraspendkey: 'penumbraspendkey',
   penumbravalid: 'penumbravalid',
   penumbrawalletid: 'penumbrawalletid',
   plpid: 'plpid',
+  penumbracompat1: 'penumbracompat1',
+  tpenumbra: 'tpenumbra',
 } as const;
 
 export type Prefix = keyof typeof Prefixes;