+++ packages/getters/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @penumbra-zone/getters
 
+## 21.0.0
+
+### Patch Changes
+
+- Updated dependencies [95d5fd9]
+  - @penumbra-zone/protobuf@7.0.0
+  - @penumbra-zone/bech32m@11.0.0
+
 ## 20.0.0
 
 ### Patch Changes