+++ packages/perspective/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @penumbra-zone/perspective
 
+## 42.0.0
+
+### Patch Changes
+
+- Updated dependencies [95d5fd9]
+  - @penumbra-zone/protobuf@7.0.0
+  - @penumbra-zone/bech32m@11.0.0
+  - @penumbra-zone/wasm@36.0.0
+  - @penumbra-zone/getters@21.0.0
+
 ## 41.0.0
 
 ### Patch Changes